Moved shared functions to clinvar_jsons_shared_lib
